relocate retry func
